Small spelling and Markdown style fixes
This fixes spelling errors in the non-policy files `README.md` and
`threat-model-notes.md` in `etc/security`. (In contrast, `irp.md`
and `threat-model.md` do not appear to contain spelling errors.)

It also fixes some minor Markdown code style inconsistencies in
`threat-model.md`. This does not affect the text of that document,
nor probably even how it is rendered.