Add script to generate test-cases for `gix-diff` to validate that diff-algorithms match (diff-slider problem)
We assume that most people would expect the output of `gix-diff` to be identical to that of `git diff`. Therefore, we want an easy way to compare the output of `gix-diff` and `git diff` on a large corpus of diffs to make sure they match.

This commit tries to make as much of that process scripted and as easily reproducible as possible.

There is a repository that contains “\[t\]ools for experimenting \[with\] diff "slider" heuristics”: [diff-slider-tools]. We can use `diff-slider-tools` to, for any git repository, generate a list of locations where a diff between two files is not unambiguous.

This commit creates a tool that takes this a list of locations generated by `diff-slider-tools` and turns it into test cases that can be run as part of `gix-diff`’s test suite.

This enables us to, whenever we want, run large-scale tests comparing `gix-diff` to `gix diff`, hopefully uncovering any edge case we might have missed in our slider heuristic and making sure we conform to our users’ expectations.

[diff-slider-tools]: https://github.com/mhagger/diff-slider-tools

1. Follow these instructions to generate a file containing sliders: https://github.com/mhagger/diff-slider-tools/blob/b59ed13d7a2a6cfe14a8f79d434b6221cc8b04dd/README.md?plain=1#L122-L146
2. Run `create-diff-cases` to create the script in `gix-diff/tests/fixtures/`. The script which will be called `make_diff_for_sliders_repo.sh`.
   ```
   # run inside `gitoxide`
   cargo run --package internal-tools -- create-diff-cases --sliders-file $DIFF_SLIDER_TOOLS/corpus/git.sliders --worktree-dir $DIFF_SLIDER_TOOLS/corpus/git.git/ --destination-dir gix-diff/tests/fixtures/
   ```
3. Run `cargo test sliders -- --nocapture` inside `gix-diff/tests` to run the actual tests.