Fix gix-credentials tests in helper::invoke when sh is not in path
Unlike on Unix-like systems, it is common for a Windows system to
have no `sh` or `sh.exe` in `PATH`. We use `gix_path::env::shell()`
in most cases to find the appropriate path. On Windows, this finds
a suitable `sh` command in the Git for Windows installation, if
applicable, and otherwise falls back to `sh.exe`. However, this was
not being done in some `gix-credentials` tests.

The affected tests use shell script credential helper fixtures that
need to be invoked in a way that forces them to be run by a shell
(without being able to pass through parameters into `gix-command`
that are guaranteed to force this). To do this, `sh ` is prepended
to the command. Since Windows need not be able to find `sh`, this
breaks on some Windows systems.

Because this specific code is in the test suite and only affects
tests, it should be acceptable to make some imperfect assumptions
about how the required concatenation can be performed. Assume:

1. That the string returned by `gix_path::env::shell()` is valid
   Unicode, or that it is acceptable for these tests to panic.

2. That the string either already has all separators as `/` (we try
   weakly to achieve this in `shell()` but don't force it), or they
   can be changed to `/` in a straightforward way that usually
   works on files in Git repositories themselves but may not always
   work (due, for example, to `\\?\` and `//?/` paths having
   different rules for the interpretation of `/` appearing later).

This is not expeted to be safe enough for general use, but it ought
to be okay for the test suite.

This commit makes those assumptions and uses

- paths for `sh`, with added conversion of `\` to `/`, and
- its script path argument, which already had that conversion,

adding POSIX shell single-quoting to each. Quoting the path to
`sh` is needed if it is under a Git installation directory like
`C:/Program Files/Git/bin`; not doing so would introduce a test
suite regression at least as severe as the test bug being fixed.
Quoting the path to its argument is done to avoid future problems
or possibly fix an existing test bug.

This technique shoud allow all `gix-credentials` tests that are
broken in the way described above to work on affected Windows
systems. But for now it is applied in only one of two places
where the code would need to be changed, so it only fixes these
tests:

    gix-credentials::credentials helper::invoke::get
    gix-credentials::credentials helper::invoke::program::path_to_helper_as_script_to_workaround_executable_bits
    gix-credentials::credentials helper::invoke::store_and_reject

It may at first seem as though adding the quoting would be enough
by itself to cause the commands to be run in a shell, because shell
quoting is being used. However, that is only enough to cause the
`command_may_be_shell_script()` call in `Program::to_command()` to
set `use_shell` to `true`. It usually does not cause the command to
be run in a shell on Windows, where `allow_manual_arg_splitting`
defaults to `true`. The manual argument splitting is able to do
both field splitting and quote removal in simple cases, so no shell
gets used.

Although it's possible to suppress manual argument splitting
indirectly by including even more characters treated specially by a
shell, including some that manual argument splitting can't handle,
to do so would be inelegant, unclear, and possibly brittle.
Therefore, this passes a pretty good path to a likely usable
`sh.exe`.