fix[pack]: Don't decompress the delta when decompressing base objects.
It was already decompressed on line 285, so attempting to decompress it
again turns the delta stream into garbage.

Unfortunately I don't know enough about the git file format to make a
test, the best I was able to do was to compare it to a reference
implementation and see what was going wrong.

Fixes #2344