Add hyperlinks where useful in the IRP
While the text of the IRP is unchanged relative to #2192, this adds
hyperlinks to some important resources. This is to make it easier
and more efficient to carry out some steps, to clarify it in case
it is used in the future by someone who hasn't previously gone
through it or a similar process, and also to make using the IRP a
bit smoother in case there's ever a stressful situation.